test: add e2e tests for `static.directory` option
